test(templates): exercise web direct routing
